+++ src/frontend/src/v2/PrivateRoom.js
@@ -25,7 +25,7 @@ class PrivateRoom extends React.Component {
                 <table class="location-input-table">
                     <tr>
                         <div class="Wrapper">
-                            <input type="text" id="room-name" class="Input-text" placeholder="이름을 알려주세요!" />
+                            <input type="text" id="user-name" class="Input-text" placeholder="이름을 알려주세요!"  onChange={this.props.handleChange}/>
                             <label for="room-name" class="Input-label">이름</label>
                         </div>
                     </tr>
@@ -35,10 +35,14 @@ class PrivateRoom extends React.Component {
                             <label for="room-name" class="Input-label">출발 위치</label>
                         </div>
                     </tr>
+                    <tr>
+                      latitude : <input type="text" size = '16' name="lat" id={0} value={this.props.myMarker.location.lat} onChange={this.props.handleChange} /><br></br>
+                      longitude : <input type="text" size = '14' name="lng" id={0} value={this.props.myMarker.location.lng} onChange={this.props.handleChange} />
+                    </tr>
                 </table>
                 <table class="search-table">
                     <td>
-                        <button  class="btn btn-submit" onClick={() => this }>이름과 출발위치 제출</button>
+                        <button  class="btn btn-submit" onClick={() => this } onClick={() => {this.props.submit()}}>이름과 출발위치 제출</button>
                     </td>
                     <td width="20"></td>
                     <td>
@@ -48,7 +52,7 @@ class PrivateRoom extends React.Component {
                 <div class="show-other-people">
                     <table class="location-input-table">
                         <tr>
-                            
+
                         </tr>
                         <tr>
